Added conditional data generation to attributes
